changed id Long->Integer because is SERIAL in database

+++ token-api/src/main/java/com/fcm_ms/token_api/repository/DeviceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fcm_ms.token_api.entity.Device;
 
-public interface DeviceRepository extends JpaRepository<Device, Long> {
+public interface DeviceRepository extends JpaRepository<Device, Integer> {
 
   Optional<Device> findByUuid(String uuid);
   void deleteByUuid(String uuid);
